Rename FieldCopyButton component to CopyButton
